Alternative approach to manage data persistance
Presentation of the idea described in the related discussion on a
controller like pattern, where recordingManager not just simply passes
through commands but rather instructs lower level services what to do,
while these lower level services are not aware of rowing states etc.
The aim of this is to create a central controller for persistent storage
by removing the duplication of instruction (getting instruction from
recordingManager as well as making own decisions based on commands).