First attempt to fix BLE bug
This commit changes the removes @abandonware/bluetooth-hci-socket
package and updates both bleno and noble to their latest version as the
hci socket library is not really necessary with the newer versions.

Furthermore, this commeit removes the concurrent central and peripheral
hack as well as switches to the default noble initialization.