Add new BLE manager and refactor BLE peripherals
As part of the migration to a new BLE library, implement a new manager
that will handle the BLE connection and peripheral management.
The new BLE manager will be responsible for opening and closing the BLE
connection and managing the BLE peripherals. The new BLE manager will
used by the BLE peripherals to communicate with the BLE devices.

Move the HRM service to the new BLE manager and refactor the HRM service
for the new API provided by the new BLE library.